refactor: unify callback events
