+++ file_processing/column_mapper.py
@@ -5,7 +5,7 @@ from typing import Dict, Iterable, List, Optional
 import pandas as pd
 from rapidfuzz import process
 
-from core.callback_controller import CallbackEvent
+from core.callback_events import CallbackEvent
 from analytics_core.callbacks.unified_callback_manager import CallbackManager
 
 REQUIRED_COLUMNS = ["person_id", "door_id", "access_result", "timestamp"]
