+++ file_processing/exporter.py
@@ -7,7 +7,7 @@ from typing import Dict
 
 import pandas as pd
 
-from core.callback_controller import CallbackEvent
+from core.callback_events import CallbackEvent
 from analytics_core.callbacks.unified_callback_manager import CallbackManager
 from core.container import get_unicode_processor
 from core.protocols import UnicodeProcessorProtocol
