+++ file_processing/readers/archive_reader.py
@@ -13,7 +13,7 @@ from .csv_reader import CSVReader
 from .excel_reader import ExcelReader
 from .fwf_reader import FWFReader
 from .json_reader import JSONReader
-from core.callback_controller import CallbackEvent
+from core.callback_events import CallbackEvent
 from analytics_core.callbacks.unified_callback_manager import CallbackManager
 from core.protocols import UnicodeProcessorProtocol
 from ..format_detector import FormatDetector
