+++ tests/callbacks/test_upload_callbacks_split.py
@@ -2,7 +2,6 @@ import types
 import pytest
 
 from dash import no_update
-from core.callback_controller import CallbackEvent
 from upload_core import UploadCore
 from services.upload.core.processor import UploadProcessingService
 from tests.fakes import (
