+++ tests/file_processing/test_column_mapper.py
@@ -2,7 +2,7 @@ import pandas as pd
 from pathlib import Path
 
 from file_processing.column_mapper import map_columns
-from core.callback_controller import CallbackEvent
+from core.callback_events import CallbackEvent
 from analytics_core.callbacks.unified_callback_manager import CallbackManager
